Authenticate the user by API and receive a response in the form of a token
The application is developed on AndroidStudio 2.2.2.
The development of the project is to create an authorization system. The
user enters the login and password and presses the button.
The completed user data is collected in the package and sent to the
server (user data is encrypted in SHA-256 format),
The server, in turn, checks the data and, if they match, sends TOKEN
back to the application, our application, if it receives a token
Then sends the user to the next screen, if not, then to the first screen
with an error.